flash-imxrt: rewrite flexspi opRead and opWrite
The purpose of rewriting the ram-text functions `flexspi_opRead` and
`flexspi_opWrite` is to reduce the size of functions relocated to the
itcm memory and to eliminate the problem of byte reading when a script
is read and the address is not necessarily aligned to 32 bit. Reading
and writeing from/to memory is performed in 8-byte chunks, which results
from the 64bit alignment of the watermark indicator.

JIRA: RTOS-605